Gemini-DeepSeek/backend/requirements.txt
# Transit Accessibility App - Backend Dependencies

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12  # Required for file uploads

# Environment & Configuration
python-dotenv==1.0.1

# AI/ML Services
google-genai==1.61.0
google-generativeai==0.8.3

# Image Processing
Pillow==11.0.0

# Data Validation
pydantic==2.10.3

# Testing (Optional - for development)
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1  # For testing FastAPI endpoints
